Allow a no-influx option
Allowing for a no-influx option will make sure that when running the smallest
Digital Ocean instance that we will not have to waste system resources on docker
to spin up influx and grafana. This should allow the box to have more memory and
cpu avaiable to this task alone and eventually stop the OS from killing the task
due to too much memory being taken from the docker-compose task.